Rename contentLength to numberOfBytes on page metadata because it's not strictly speaking the content length (since it doesn't include entity headers)

+++ src/main/java/io/datareplication/producer/feed/FeedEntityRepository.java
@@ -21,7 +21,7 @@ public interface FeedEntityRepository {
         @NonNull ContentId contentId;
         @NonNull Timestamp lastModified;
         @NonNull Optional<@NonNull Timestamp> originalLastModified;
-        int contentLength;
+        long contentLength;
         @NonNull Optional<@NonNull PageId> pageId;
     }
